TIP-559: Move ZipFilesystemFactory in Archiver folder
