+++ src/Pim/Component/Connector/Archiver/ArchivableFileWriterArchiver.php
@@ -7,7 +7,6 @@ use Akeneo\Component\Batch\Job\JobRegistry;
 use Akeneo\Component\Batch\Model\JobExecution;
 use Akeneo\Component\Batch\Step\ItemStep;
 use League\Flysystem\Filesystem;
-use Pim\Component\Connector\Filesystem\ZipFilesystemFactory;
 use Pim\Component\Connector\Writer\File\ArchivableWriterInterface;
 
 /**
