+++ src/Pim/Component/Connector/spec/Archiver/ArchivableFileWriterArchiverSpec.php
@@ -12,7 +12,7 @@ use Akeneo\Component\Batch\Step\ItemStep;
 use League\Flysystem\Adapter\Local as LocalAdapter;
 use League\Flysystem\Filesystem;
 use PhpSpec\ObjectBehavior;
-use Pim\Component\Connector\Filesystem\ZipFilesystemFactory;
+use Pim\Component\Connector\Archiver\ZipFilesystemFactory;
 use Pim\Component\Connector\Writer\File\Csv\Writer;
 use Prophecy\Argument;
